+++ src/utils/enum.js
@@ -1,0 +1,57 @@
+export const TimeValue = {
+  MILLISECONDS_IN_SECOND: 1000,
+  SECONDS_IN_MINUTE: 60,
+  MINUTES_IN_HOUR: 60,
+  HOURS_IN_DAY: 24,
+  MILLISECONDS_IN_MINUTE: 1000 * 60,
+  MILLISECONDS_IN_HOUR: 1000 * 60 * 60,
+  MILLISECONDS_IN_DAY: 1000 * 60 * 60 * 24,
+};
+
+export const Position = {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`,
+  AFTEREND: `afterend`
+};
+
+export const SortType = {
+  TIME: `time`,
+  PRICE: `price`,
+  DEFAULT: `default`
+};
+
+export const FilterType = {
+  EVERYTHING: `everything`,
+  FUTURE: `future`,
+  PAST: `past`
+};
+
+export const Action = {
+  CREATE: `create`,
+  UPDATE: `update`,
+  DELETE: `delete`,
+  NONE: `none`
+};
+
+export const Mode = {
+  ADDING: `adding`,
+  DEFAULT: `default`,
+};
+
+export const Method = {
+  GET: `GET`,
+  POST: `POST`,
+  PUT: `PUT`,
+  DELETE: `DELETE`
+};
+
+export const Key = {
+  ESCAPE: `Escape`,
+  IE_ESC: `Esc`,
+  BACKSPACE: `Backspace`,
+  DELETE: `Delete`
+};
+
+export const TagName = {
+  INPUT: `INPUT`
+};
